Added controllers for user
